requirements.txt
# Core dependencies
rasa>=3.6.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0

# HTTP client for external APIs (MISSING - ADDED)
httpx>=0.24.0
aiofiles>=0.8.0

# Messaging integrations
twilio>=7.0.0
requests>=2.26.0

# Utilities
python-dotenv>=0.19.0
pyjwt>=2.3.0
cryptography>=3.4.0
passlib>=1.7.4
tenacity>=8.0.0
schedule>=1.1.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
httpx>=0.18.0

# Monitoring
prometheus-client>=0.11.0
grafana-api>=1.0.0

# Additional health-specific packages (MISSING - ADDED)
python-dateutil>=2.8.0
pytz>=2021.3
